helpers/content: respect <!-- more --> tag when truncating content
This is arianf/Ghost@57f1daba05249303bb5a0be9119b9f572c093951 tweaked to match
on <!-- more --> (extra spaces) as my content got imported from tumblr, which
uses that rather than <!--more-->

There was an issue for this opened upstream (TryGhost/Ghost#4933), but got
rejected by the developers. While I understand the objections raised by the
designer from the point of view of the visual artist, this really ignores the
writer as the literary artist; the 26 word excerpts might work well visually,
but they more often than not do not work well in textual terms, so one has to
decide whether the visual or literary aesthetic should take a priority -- in
my case, that is a simple choice, the literary takes precedence.

This commit, like the original it is based on, modifies the 'content' token,
not the excerpt, and it needs to be accompanied by an appropriate theme that
uses {{content}} rather than {excerpt}} for the index page loop (I use a
modified Casper theme).